Reland "Upconvert various types to int.", misc. codecs portion.
This reverts portions of commit cb180976dd0e9672cde4523d87b5f4857478b5e9, which
reverted commit 83ad33a8aed1fb00e422b6abd33c3e8942821c24.  Specifically, the
files in webrtc/modules/audio_coding/codecs/ that are not in ilbc/ or isac/, as
well as webrtc/modules/audio_coding/main/test/opus_test.cc, are relanded.

The original commit message is below:

Upconvert various types to int.

Per comments from HL/kwiberg on https://webrtc-codereview.appspot.com/42569004 , when there is existing usage of mixed types (int16_t, int, etc.), we'd prefer to standardize on larger types like int and phase out use of int16_t.

Specifically, "Using int16 just because we're sure all reasonable values will fit in 16 bits isn't usually meaningful in C."

This converts some existing uses of int16_t (and, in a few cases, other types such as uint16_t) to int (or, in a few places, int32_t).  Other locations will be converted to size_t in a separate change.

BUG=none
TBR=kwiberg

Review URL: https://codereview.webrtc.org/1179093003

Cr-Commit-Position: refs/heads/master@{#9424}

+++ webrtc/modules/audio_coding/codecs/cng/include/webrtc_cng.h
@@ -68,8 +68,8 @@ int16_t WebRtcCng_CreateDec(CNG_dec_inst** cng_inst);
  *                      -1 - Error
  */
 
-int16_t WebRtcCng_InitEnc(CNG_enc_inst* cng_inst, uint16_t fs, int16_t interval,
-                          int16_t quality);
+int WebRtcCng_InitEnc(CNG_enc_inst* cng_inst, int fs, int16_t interval,
+                      int16_t quality);
 int16_t WebRtcCng_InitDec(CNG_dec_inst* cng_inst);
 
 /****************************************************************************
@@ -103,9 +103,9 @@ int16_t WebRtcCng_FreeDec(CNG_dec_inst* cng_inst);
  * Return value       :  0 - Ok
  *                      -1 - Error
  */
-int16_t WebRtcCng_Encode(CNG_enc_inst* cng_inst, int16_t* speech,
-                         int16_t nrOfSamples, uint8_t* SIDdata,
-                         int16_t* bytesOut, int16_t forceSID);
+int WebRtcCng_Encode(CNG_enc_inst* cng_inst, int16_t* speech,
+                     int16_t nrOfSamples, uint8_t* SIDdata,
+                     int16_t* bytesOut, int16_t forceSID);
 
 /****************************************************************************
  * WebRtcCng_UpdateSid(...)
